requirements.txt
Django==4.1
djangorestframework==3.13.1
supervisor==4.1.0
requests
